requirements.txt
apturl==0.5.2
argcomplete==1.8.1
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.3
Babel==2.4.0
beautifulsoup4==4.6.0
bleach==2.1.2
Brlapi==0.6.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
entrypoints==0.2.3.post1
fake-useragent==0.1.11
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
instaloader==4.2.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.16
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
jwt==0.6.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
pandas==0.24.2
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
progressbar==2.3
prompt-toolkit==1.0.15
protobuf==3.0.0
psutil==5.4.2
pyasn1==0.1.9
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.6.1
python-debian==0.1.32
python-gnupg==0.4.1
python-xlib==0.20
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
rsa==3.3
safeeyes==2.0.9
scour==0.36
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
wrapt==1.11.1
xkit==0.0.0
xlrd==1.2.0
zope.interface==4.3.2
